requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
Django==3.1.6
idna==2.10
psycopg2==2.8.6
pycparser==2.20
python-telegram-bot==13.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
